requirements.txt
absl-py==0.8.1
cycler==0.10.0
decorator==4.4.0
future==0.18.1
grpcio==1.24.3
imageio==2.6.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
packaging==19.2
pandas==0.25.2
Pillow==6.2.1
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytorch-lightning==0.5.2.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.1
six==1.12.0
tb-nightly==1.15.0a20190708
test-tube==0.7.3
torch==1.3.0
torchvision==0.4.1
tqdm==4.36.1
Werkzeug==0.16.0